+++ dwarfeng-ftp-api/src/test/java/com/dwarfeng/ftp/api/integration/example/TelqosExample.java
@@ -31,8 +31,6 @@ public class TelqosExample {
 
         System.out.println("示例演示完毕, 感谢您测试与使用!");
 
-        ctx.stop();
-        ctx.close();
         System.exit(0);
     }
 }