+++ dwarfeng-ftp-core/src/test/java/com/dwarfeng/ftp/node/example/ClearDirectoryExample.java
@@ -65,8 +65,6 @@ public class ClearDirectoryExample {
         ftpHandler.removeDirectory(new String[]{rootPath, middlePath});
         System.out.println("示例演示完毕, 感谢您测试与使用!");
 
-        ctx.stop();
-        ctx.close();
         System.exit(0);
     }
 }