+++ dwarfeng-ftp-core/src/test/java/com/dwarfeng/ftp/node/example/CopyFileExample.java
@@ -91,8 +91,6 @@ public class CopyFileExample {
         ftpHandler.removeDirectory(new String[]{rootPath, "漫画女孩-1"});
         System.out.println("示例演示完毕, 感谢您测试与使用!");
 
-        ctx.stop();
-        ctx.close();
         System.exit(0);
     }
 }