+++ dwarfeng-ftp-core/src/test/java/com/dwarfeng/ftp/node/example/DescFileExample.java
@@ -82,8 +82,6 @@ public class DescFileExample {
         ftpHandler.removeDirectory(new String[]{rootPath, "漫画女孩"});
         System.out.println("示例演示完毕, 感谢您测试与使用!");
 
-        ctx.stop();
-        ctx.close();
         System.exit(0);
     }
 }