+++ dwarfeng-ftp-core/src/test/java/com/dwarfeng/ftp/node/example/ListFileExample.java
@@ -77,8 +77,6 @@ public class ListFileExample {
         ftpHandler.deleteFile(new String[]{rootPath}, "漫画女孩.jpg");
         System.out.println("示例演示完毕, 感谢您测试与使用!");
 
-        ctx.stop();
-        ctx.close();
         System.exit(0);
     }
 }