+++ dwarfeng-ftp-core/src/test/java/com/dwarfeng/ftp/node/example/StreamExample.java
@@ -138,8 +138,6 @@ public class StreamExample {
         ftpHandler.removeDirectory(new String[]{rootPath, "漫画女孩"});
         System.out.println("示例演示完毕, 感谢您测试与使用!");
 
-        ctx.stop();
-        ctx.close();
         System.exit(0);
     }
 }